move YouTubeVideo.js to folder
