+++ components/OneDay/Hero.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { above, below } from '../../utilities';
 import HeroContainer from '../shared/HeroContainer';
-import YouTubeVideo from '../shared/YouTubeVideo';
+import YouTubeVideo from '../shared/YouTubeVideo/YouTubeVideo';
 import LinkButton from '../shared/LinkButton/LinkButton';
 
 const VIDEO_ID = 'cAftppcmqRE';
