+++ pages/wi/2021/tickets.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { NextSeo } from 'next-seo';
 
 import ContentSection from '../../../components/shared/ContentSection';
-import YouTubeVideo from '../../../components/shared/YouTubeVideo';
+import YouTubeVideo from '../../../components/shared/YouTubeVideo/YouTubeVideo';
 import LinkButton from '../../../components/shared/LinkButton/LinkButton';
 import { ActionButtonRow } from '../../../components/shared/StandardStyles';
 import { above, below } from '../../../utilities';
